Fixed logic bug in uri.load_resource and updated modelspec.load to use uri
Changed ordering of if/else in load_resource to only load json
if actually present. Previous "if r.json" statement always returned
true because it was checking fo the existence of the response
object's .json method intead of the content of r.json(), which
returns a JSONdecode error if no json data is present.

Updated modelspec.load_modelspec() to pass through to
uri.load_resource for convenience since the latter allows
both local files and http/etc.