WIP: Added support for signal segments
This commit adds support to the RasterizedSignal class for supporting
data segments. Note that I did not make this a RasterizedSignalSubset
class since it was actually pretty straightforward to add some support
to the main class. This was an extension of the original idea to add a
`t0` property to indicate when the data begins relative to recording
start, but that has now been superseded by a new data structure,
segments, which indicates the start/stop times of the various data
segments. For the default RasterizedSignal (i.e., when we have not
selected any subsets, the data is treated as a list of subsets
containing only one subset).