Added wc_idx and fir_idx options to strf_heatmap
Instead of using the first instance of wc and fir by default
for plotting, strf_heatmap will now use the idxth+1 th instance
of that module. For example, wc_idx=1 will use the second
instance of weight_channels in the modelspec.

Added dynamic_sigmoid to quickplot which takes advantage
of the above option.