Fix bug introduced by revamp of model registry (#151)
New import system for files broke the JSONification of the registry.
Fixed this by adding a source_string to the Keyword object that's used
to identify how to locate the keyword. This allows us to preserve the
original method of finding the keyword (e.g., via an import or by
specifying the file).