Added basic_with_copy cost function
Added a new cost function that is mostly a duplicate of
basic_cost. The one difference is the kwarg "copy_phi",
which takes in a list of 2-tuples to specify a pattern of
copying phi from the first entry to the fn_kwargs of the second
entry. This was added to solve a problem with a particular model
in which one module needed reference to the parameters of another
module during evaluation. It seemed possible, however, that other
models might run into the same need so I've added it as a more
general-use function.

Example: If I wanted to let the module at index 4 have reference
to the parameters from the modules at indices 1 and 3, I would
specify

	copy_phi=[(1, 4), (3, 4)]

To get the argument into the cost_function, one way is to define
a wrapper inside whatever fitting function is being used.
For example:

	def my_cost(*args, **kwargs):
		c = [(1, 4), (3, 4)]
		return basic_with_copy(*args, **kwargs, copy_phi=c)