Added ModelSpec.drop_module method, basic_error
ModelSpec.drop_module either removes a module in-place
or returns a new ModelSpec that does not contain the
dropped module, using mod_idx to select it. Intended
use is for easily comparing model evaluations
with and without a particular module included.

basic_error takes in many of the same arguments as
fit_basic, but just evaluates the model and returns
the error instead of fitting.