Updated xforms.load_recordings and ld loader keyword
load_recordings now accepts two new cellid formats:
1) siteids: If no hyphen is present in cellid, it will be
interpreted as a siteid. In that case, all channels that contain
the siteid will be extracted.
2) None:  if cellid=None, all channels will be extracted.

Also fixed generate_stim_from_epochs so that it will not cause
an error if no matches are found for epoch_regex. Previously
this effectively hardcoded a requirement that all recordings contain
"TRIAL" epochs when using load_recordings.