Updated previous change to fit_model_xform behavior
initial_context is now used to update xforms_init_context
which is then passed to generate_xform_spec. The result is
that anything stored in initial_context is also added to
the keyword arguments of the xforms registry and therefore
stored along with the xfspec as desired. The previous change
just added initial_context directly to context argument of
xforms.evaluate, which does not get saved.