Added evaluate_tf to modelspec.py
Added a new evaluate_tf function to modelspec.py for evaluating
tensorflow layers that are tagged as 'tf_only' in their 'fn' entry.
Modelspec.evalute will not call evaluate_tf instead of evaluate
when at least one module is tagged in that way.

TODO: Currently using the flag in the 'fn' name as a kludge, but
it might cause issues with the gui browser or other functions that
rely on looking up the function designated by 'fn.' Will work on
a better solution long-term.