Implemented Conv2D for NEMS
New TF layers:
-Conv2D_NEMS: As tf.keras.layers.Conv2D.
-Dense_NEMS: As tf.keras.layers.Dense.
-WeightChannelsNew: As nems.tf.layers.WeightChannelsBasic but
implemented as a weighted sum instead of conv1d and compatible with
Conv2D output format.
-Note: Each of the above layers is flagged as _TF_ONLY for evaluations.
In the short-term this means some extra checks had to be added to
cnnlink_new functions and BaseLayer to avoid conflicts. For example,
these layers will not work with use_modelspec_init for now. However,
these extra checks should not interfere with the existing model
architectures.

New keywords:
-conv2d, dense, wcn for the above layers, respectively.
-Note: as with the layers, these keywords are tagged as 'tf_only' in
their 'fn' entry. This is a temporary kludge to signal to a Modelspec
object that it should use evaluate_tf instead of evaluate.